FIR/UAST: commonize binary expression (bitwise operator)
